syscall: add Flock_t.Lock method
Fixes #7059

R=golang-codereviews, iant, mikioh.mikioh
CC=golang-codereviews
https://golang.org/cl/53470043

+++ src/pkg/syscall/consistency_unix_test.go
@@ -37,8 +37,8 @@ func _() {
 	_ = syscall.Flock_t{
 		Type:   int16(0),
 		Whence: int16(0),
-		Start:  0,
-		Len:    0,
+		Start:  int64(0),
+		Len:    int64(0),
 		Pid:    int32(0),
 	}
 }
